Revert "feat(builder): enable minifyJS/minifyCSS of  HTML plugin by default (#4433)"
This reverts commit 31b6d0088a95cb851bae31833a96653e50d2eb3f.
