--- .changeset/many-walls-greet.md
@@ -1,7 +1,0 @@
----
-'@modern-js/builder-shared': patch
----
-
-chore(builder): update htmlPlugin default value(enable minifyJS / minifyCSS / ... by default)
-
-chore(builder): 更新 htmlPlugin 默认值 (默认开启 minifyJS、minifyCSS 等能力)
