+++ packages/builder/builder-shared/package.json
@@ -145,7 +145,6 @@
     "@types/node": "^14",
     "html-webpack-plugin": "5.5.3",
     "terser": "5.19.2",
-    "terser-webpack-plugin": "5.3.9",
     "typescript": "^5"
   },
   "publishConfig": {
