+++ packages/builder/builder-webpack-provider/src/index.ts
@@ -38,4 +38,5 @@ export type {
   WebpackConfig,
   CSSExtractOptions,
   HTMLPluginOptions,
+  TerserPluginOptions,
 } from './types';
