+++ packages/builder/builder-webpack-provider/src/types/config/tools.ts
@@ -6,7 +6,6 @@ import type {
   SharedToolsConfig,
   ToolsLessConfig,
   ToolsSassConfig,
-  TerserPluginOptions,
 } from '@modern-js/builder-shared';
 import type {
   ModifyWebpackChainUtils,
@@ -17,6 +16,7 @@ import type {
   CssMinimizerPluginOptions,
   HTMLPluginOptions,
   InspectorPluginOptions,
+  TerserPluginOptions,
   TSLoaderOptions,
   WebpackChain,
   WebpackConfig,
