+++ packages/document/builder-doc/docs/en/config/tools/htmlPlugin.md
@@ -9,18 +9,14 @@ const defaultHtmlPluginOptions = {
   template, // defaults to the built-in HTML template path
   templateParameters, // corresponding to the html.templateParameters config
   chunks: [entryName],
-  minify: { // generated based on output.disableMinimize
-    removeComments: true,
+  minify: {
+    removeComments: false,
     useShortDoctype: true,
     keepClosingSlash: true,
     collapseWhitespace: true,
     removeRedundantAttributes: true,
     removeScriptTypeAttributes: true,
     removeStyleLinkTypeAttributes: true,
-    removeEmptyAttributes: true,
-    minifyJS, // generated based on output.charset and tools.terser.terserOptions
-    minifyCSS: true,
-    minifyURLs: true,
   },
 };
 ```
@@ -69,25 +65,6 @@ The built-in `html-webpack-plugin` plugins can be disabled by set `tools.htmlPlu
 export default {
   tools: {
     htmlPlugin: false,
-  },
-};
-```
-
-### Disable JS/CSS minify
-
-By default, Builder will compresses JavaScript/CSS code inside HTML during the production build to improve the page performance. This ability is often helpful when using custom templates or inserting custom scripts.
-
-However, when `output.enableInlineScripts` or `output.enableInlineStyles` is turned on, inline JavaScript/CSS code will be repeatedly compressed, which will have a certain impact on build performance. You can modify the default minify behavior by modifying the `tools.htmlPlugin.minify` configuration.
-
-```js
-export default {
-  tools: {
-    htmlPlugin: config => {
-      if (typeof config.minify === 'object') {
-        config.minify.minifyJS = false;
-        config.minify.minifyCSS = false;
-      }
-    },
   },
 };
 ```
