--- tests/e2e/builder/cases/html/minify/src/App.css
@@ -1,16 +1,0 @@
-html,
-body {
-  padding: 0;
-  margin: 0;
-}
-
-* {
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  box-sizing: border-box;
-}
-
-.test {
-  text-align: center;
-  line-height: 1.5;
-}
