--- tests/e2e/builder/cases/html/minify/src/App.tsx
@@ -1,9 +1,0 @@
-import './App.css';
-
-const App = () => (
-  <div id="test" className="test">
-    Hello Builder!
-  </div>
-);
-
-export default App;
