--- tests/e2e/builder/cases/html/minify/src/index.ts
@@ -1,5 +1,0 @@
-import React from 'react';
-import { render } from 'react-dom';
-import App from './App';
-
-render(React.createElement(App), document.getElementById('root'));
