--- tests/e2e/builder/cases/html/minify/static/index.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>custom template</title>
-    <script>
-     window.a = 1;
-     window.b = 2;
-     console.log(111111);
-     console.info(111111);
-     console.warn(111111);
-    </script>
-    <style>
-      .test {
-        font-size: 146px;
-        background-color: green;
-      }
-
-      #a {
-        text-align: center;
-        line-height: 1.5;
-        font-size: 1.5rem;
-      }
-
-      #b {
-        text-align: center;
-        line-height: 1.5;
-        font-size: 1.5rem;
-        background: #fafafa;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="test-template">bar</div>
-    <div id="root" />
-  </body>
-</html>
