--- tests/e2e/builder/cases/html/minify/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "@modern-js/tsconfig/base",
-  "compilerOptions": {
-    "declaration": false,
-    "jsx": "react-jsx",
-    "baseUrl": "./",
-    "outDir": "./dist",
-    "paths": {
-      "@scripts/*": ["../../../scripts/*"]
-    }
-  },
-  "include": ["src"]
-}
